docs-next: redo <Sheet> component signature
In the orignal PR for the new @threlte/theatre docs, the <Sheet>
component signature was erroniously identical to that of <Project>.

This fixes the discrepency and adds the correct signature.

Note: some types seemed overly complex and were simplified for the
sake of brevity, most notably that of the play binding. Since
the description field doesn't seem to work for bindings, a note
was added inside the type field to point the reader to the
Theatre.js docs for more information.

In the future, this can be improved.