Your First Scene 2.0
This rewrites the `Your First Scene` tutorial to make it more friendly
to beginners. Instead of focusing on Threlte technical features,
sections are titled after the changes they achieve in the scene.
Additionally, the progress is shown throw examples in every step,
allowing the new user to see the changes as they occur.