Fix SequenceController autoplay and delay bugs
This fixes bugs relating to sequence autoplay and delay logic. The
logic was erroneosly placed in the config method rather than the
constructor. In order to allow triggering autoplay in the <Sequence>
component, an autoplay method is also exposed.

To verify this logic as well as on-the-fly reconfiguration, the
sequence example is augmented with new props.